Separate Controllers Fixes #18
