+++ routes/rIndex.js
@@ -1,5 +1,5 @@
 import express from 'express';
-import results from './results.js';
+import results from './rResults.js';
 
 const router = express.Router();
